refactor: expose sha1 via common helper
