+++ scripts/common.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const crypto = require("crypto");
 
-module.exports = function sha1(file) {
+function sha1(file) {
   return new Promise((resolve, reject) => {
     const hash = crypto.createHash("sha1");
     const stream = fs.createReadStream(file);
@@ -9,4 +9,8 @@ module.exports = function sha1(file) {
     stream.on("data", (chunk) => hash.update(chunk));
     stream.on("end", () => resolve(hash.digest("hex")));
   });
+}
+
+module.exports = {
+  sha1,
 };
